Re-struct ci_buf interface to use ci_buffers/mem pools
Also fix it to use correct integer type for memory sizes (size_t instead
of int).

TODO: ci_buf should be used only internally by c-icap library and c-icap
server. Also should be renamed to not confused with the ci_buffer API.

+++ txt_format.c
@@ -637,7 +637,7 @@ int fmt_req_preview_len(ci_request_t *req, char *buf, int len, const char *param
         return 0;
 
     if (req->preview >= 0)
-        return snprintf(buf, len, "%d", req->preview_data.used);
+        return snprintf(buf, len, "%ld", (long)req->preview_data.used);
 
     *buf = '-';
     return 1;
